requirements.txt
Flask==2.3.2
pymongo==4.5.0
pandas==2.0.3
plotly==5.15.0
reportlab==4.0.4
python-dotenv==1.0.0
Werkzeug==2.3.6
